refactor: allow import of docling datamodel without transformers
